Two steps in goal definition
